property_no（物件番号）へ挿入する値がIntegerの範囲を超えているため DecimalField(max_digits=10, decimal_places=0, null=True)に修正

+++ landinfo/search/models.py
@@ -50,7 +50,7 @@ class LandInfo(models.Model):
     estab_completion_time = models.CharField(max_length=50, null=True)  # 造成完成時期
     delivery = models.CharField(max_length=50, null=True)  # 引渡し（時期/方法）
     scheduled_sales = models.CharField(max_length=200, null=True)  # 販売スケジュール
-    property_no = models.IntegerField(null=True)  # 物件番号
+    property_no = models.DecimalField(max_digits=10, decimal_places=0, null=True)  # 物件番号
     contact_infomation = models.CharField(max_length=200, null=True)  # お問い合わせ先
     seller = models.CharField(max_length=200, null=True)  # 売主
     info_provider = models.CharField(max_length=200, null=True)  # 情報提供元
